Combine vehicle controllers
