+++ src/main/java/evswap/swp391to4/controller/VehicleController.java
@@ -1,19 +1,25 @@
 package evswap.swp391to4.controller;
 
+import evswap.swp391to4.dto.VehicleRegistrationForm;
+import evswap.swp391to4.entity.Driver;
 import evswap.swp391to4.entity.Vehicle;
+import evswap.swp391to4.service.DriverService;
 import evswap.swp391to4.service.VehicleService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
-@RestController
+@Controller
 @RequiredArgsConstructor
-@RequestMapping("/api/drivers/{driverId}/vehicles")
 public class VehicleController {
 
+    private final DriverService driverService;
     private final VehicleService vehicleService;
 
-    @PostMapping
+    @PostMapping("/api/drivers/{driverId}/vehicles")
     public ResponseEntity<Vehicle> addVehicle(@PathVariable Integer driverId,
                                               @RequestBody VehicleRequest request) {
         Vehicle vehicle = Vehicle.builder()
@@ -23,6 +29,61 @@ public class VehicleController {
                 .build();
         Vehicle savedVehicle = vehicleService.addVehicleToDriver(driverId, vehicle);
         return ResponseEntity.ok(savedVehicle);
+    }
+
+    @GetMapping("/vehicles/register")
+    public String showRegistrationForm(@RequestParam("driverId") Integer driverId,
+                                       Model model,
+                                       RedirectAttributes redirect) {
+        try {
+            Driver driver = driverService.getDriverById(driverId);
+
+            model.addAttribute("driverId", driver.getDriverId());
+            model.addAttribute("driverName", driver.getFullName());
+            model.addAttribute("driverInitial", extractInitial(driver.getFullName()));
+
+            if (!model.containsAttribute("vehicleForm")) {
+                model.addAttribute("vehicleForm", new VehicleRegistrationForm());
+            }
+
+            return "vehicle-register";
+        } catch (Exception e) {
+            redirect.addFlashAttribute("loginError", e.getMessage());
+            return "redirect:/login";
+        }
+    }
+
+    @PostMapping("/vehicles/register")
+    public String registerVehicle(@RequestParam("driverId") Integer driverId,
+                                  @ModelAttribute("vehicleForm") VehicleRegistrationForm form,
+                                  RedirectAttributes redirect) {
+        try {
+            Vehicle vehicle = Vehicle.builder()
+                    .model(form.getModel())
+                    .vin(form.getVin())
+                    .plateNumber(form.getPlateNumber())
+                    .build();
+
+            vehicleService.addVehicleToDriver(driverId, vehicle);
+            redirect.addFlashAttribute("loginSuccess", "Đăng ký phương tiện thành công! Vui lòng đăng nhập.");
+            return "redirect:/login";
+        } catch (Exception e) {
+            redirect.addFlashAttribute("vehicleError", e.getMessage());
+            redirect.addFlashAttribute("vehicleForm", form);
+            redirect.addAttribute("driverId", driverId);
+            return "redirect:/vehicles/register";
+        }
+    }
+
+    private String extractInitial(String fullName) {
+        if (fullName == null) {
+            return "U";
+        }
+        String trimmed = fullName.trim();
+        if (trimmed.isEmpty()) {
+            return "U";
+        }
+        return trimmed.substring(0, 1).toUpperCase();
     }
 
     public record VehicleRequest(String vin, String plateNumber, String model) {
